src/util/NumberUtil.ts
export class NumberUtil {

    public static isPrime(n: number): boolean {
        if (n === undefined) {
            return false;
        }
        if (n <= 1) {
            return false;
        }
        if (n <= 3) {
            return true;
        }
        if (n % 2 === 0 || n % 3 === 0) {
            return false;
        }
        let i = 5;
        while (i * i <= n) {
            if (n % i === 0 || n % (i + 2) === 0) {
                return false;
            }
            i = i + 6;
        }
        return true;
    }

}
